refactor: move criticalhits command creation to the calculator
